restructuring and adding comment for future
