+++ client/src/containers/QuizContainers/QuestionContainer.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import QuizQuestion from '../../presentation/Dashboard/Quiz/QuizBegin/QuizQuestion/QuizQuestion';
-import addUserAnswer from '../../actionCreators/sync/quiz/addUserAnswer';
+import addUserAnswer from '../../actionCreators/sync/quiz/addUserAnswer/addUserAnswer';
 
 function mapStateToProps(state){
   return {
